requirements.txt
ipywidgets==8.0.4
indic-transliteration==2.3.3
googletrans==4.0.0-rc1
google-api-python-client==2.122.0
tqdm==4.65.0
requests==2.32.2
youtube-transcript-api==0.6.2
yt-dlp==2024.7.9
python-dotenv==1.0.1
pytube==15.0.0
pydub==0.25.1
numpy==1.23.5
pandas==1.4.0
torch
transformers
tensorflow==2.13.0rc2
